+++ spiel/__init__.py
@@ -2,6 +2,18 @@ from spiel.app import SuspendType, present
 from spiel.constants import __version__
 from spiel.deck import Deck
 from spiel.slide import Slide
+from spiel.transitions.protocol import Direction, Transition
+from spiel.transitions.swipe import Swipe
 from spiel.triggers import Triggers
 
-__all__ = ["present", "SuspendType", "__version__", "Deck", "Slide", "Triggers"]
+__all__ = [
+    "Deck",
+    "Direction",
+    "Slide",
+    "SuspendType",
+    "Swipe",
+    "Transition",
+    "Triggers",
+    "__version__",
+    "present",
+]