+++ spiel/widgets/fixed_slide.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
 
 import sys
-from time import monotonic
 
 from rich.box import HEAVY
 from rich.console import RenderableType
@@ -10,32 +9,25 @@ from rich.panel import Panel
 from rich.protocol import is_renderable
 from rich.style import Style
 from rich.traceback import Traceback
-from textual.reactive import reactive
 
 import spiel
 from spiel.exceptions import SpielException
+from spiel.slide import Slide
 from spiel.triggers import Triggers
 from spiel.widgets.widget import SpielWidget
 
 
-class SlideWidget(SpielWidget):
-    triggers: Triggers = reactive(Triggers.new)  # type: ignore[assignment,arg-type]
+class FixedSlideWidget(SpielWidget):
+    def __init__(self, slide: Slide, triggers: Triggers | None = None, id: str | None = None):
+        super().__init__(id=id)
 
-    def on_mount(self) -> None:
-        super().on_mount()
-
-        if not self.app.fixed_triggers:
-            self.set_interval(self.app.slide_refresh_rate, self.update_triggers)
-        else:
-            self.triggers = self.app.fixed_triggers
-
-    def update_triggers(self) -> None:
-        self.triggers = Triggers(now=monotonic(), _times=self.triggers._times)
+        self.slide = slide
+        self.triggers = triggers or Triggers.new()
 
     def render(self) -> RenderableType:
         try:
             self.remove_class("error")
-            r = self.current_slide.render(triggers=self.triggers)
+            r = self.slide.render(triggers=self.triggers)
             if is_renderable(r):
                 return r
             else:
@@ -43,7 +35,7 @@ class SlideWidget(SpielWidget):
         except Exception:
             self.add_class("error")
             et, ev, tr = sys.exc_info()
-            if et is None or ev is None or tr is None:  # pragma: unreachable
+            if et is None or ev is None or tr is None:
                 raise SpielException("Expected to be handling an exception, but wasn't.")
             return Panel(
                 Traceback.from_exception(